pkf_snark_executive/requirements.txt
streamlit>=1.30.0
ezdxf>=0.18.0
pdfplumber>=0.10.0
reportlab>=4.0.0
openpyxl>=3.1.0
plotly>=5.18.0
httpx>=0.25.0
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
